tests/requirements.txt
click == 8.1.7
PySide6 == 6.7.1
pytest == 8.2.1
mayatest == 0.1.2
wemake-python-styleguide == 0.19.2